Restructuring the solution into different projects, Server and Library, for better organization
