+++ Library/Models/Acceleration.cs
@@ -1,0 +1,9 @@
+﻿namespace backend.Library.Models
+{
+    public readonly struct Acceleration
+    {
+        public float X { init; get; }
+        public float Y { init; get; }
+        public float Z { init; get; }
+    }
+}
