+++ Library/Services/DataProcessors/DataExtractors/TemperatureExtractor.cs
@@ -1,0 +1,24 @@
+﻿using backend.Library.Services;
+using backend.Library.Services.DataProviders;
+using Microsoft.Extensions.DependencyInjection;
+
+namespace backend.Library.Services.DataProcessors.DataExtractors
+{
+    /// <summary>
+    /// Extracts temperature data from SerialProvider data.
+    /// </summary>
+    public class TemperatureExtractor : DataExtractorBase<float>
+    {
+        public TemperatureExtractor(
+            [FromKeyedServices(ServiceKeys.SerialProvider)]
+                IDataProvider<Dictionary<SerialProvider.DataLabel, string>> provider
+        )
+            : base(provider)
+        {
+            // Extract temperature data
+            _sourceIndexes = [SerialProvider.DataLabel.Temperature];
+        }
+
+        protected override float Convert(IEnumerable<string> data) => float.Parse(data.First());
+    }
+}
