+++ Library/Services/DataProviders/IDataProvider.cs
@@ -1,0 +1,17 @@
+﻿using backend.Library.Models;
+
+namespace backend.Library.Services.DataProviders
+{
+    /// <summary>
+    /// Contract for a data provider.
+    /// </summary>
+    /// <typeparam name="T">The type of the data provided</typeparam>
+    public interface IDataProvider<T>
+    {
+        /// <summary>
+        /// Event triggered whenever the provider has new data.
+        /// Subscribe to this event to act on the new data.
+        /// </summary>
+        public event Action<EventData<T>>? OnDataProvided;
+    }
+}
