+++ Library/Services/ServiceKeys.cs
@@ -1,0 +1,13 @@
+﻿namespace backend.Library.Services
+{
+    public enum ServiceKeys
+    {
+        SerialProvider,
+        PressureExtractor,
+        TemperatureExtractor,
+        AltitudeExtractor,
+        AltitudeGPSExtractor,
+        AltitudeDeltaProcessor,
+        VelocityProcessor,
+    }
+}
