--- Models/Acceleration.cs
@@ -1,9 +1,0 @@
-﻿namespace backend24.Models
-{
-    public readonly struct Acceleration
-    {
-        public float X { init; get; }
-        public float Y { init; get; }
-        public float Z { init; get; }
-    }
-}
