--- Services/ServiceKeys.cs
@@ -1,13 +1,0 @@
-﻿namespace backend24.Services
-{
-    public enum ServiceKeys
-    {
-        SerialProvider,
-        PressureExtractor,
-        TemperatureExtractor,
-        AltitudeExtractor,
-        AltitudeGPSExtractor,
-        AltitudeDeltaProcessor,
-        VelocityProcessor,
-    }
-}
